@@ -1,12 +1,10 @@
 package jrizani.jrspinner.example;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
-
-import java.util.List;
 
 import jrizani.jrspinner.JRSpinner;
 
